git-svn-id: https://subversion.assembla.com/svn/mediasl/Projectes%20MPLAB/EOS/trunk@6317 4ad52a4b-eac5-49b1-a5a4-524d969995ed

+++ v2/Demo/FormsDemo/Main.cpp
@@ -155,7 +155,7 @@ void MyApplication::selectorNotifyEventHandler(
     if (formsService != nullptr) {
         FormHandle form = formsService->getActiveForm();
         if (form != nullptr) {
-            int16_t delta = notification.position - oldPosition;
+            int delta = notification.position - oldPosition;
             if (delta != 0) {
                 message.id = MSG_SELECTOR;
                 message.target = form;
